+++ engine/src/graphics.cpp
@@ -1,4 +1,4 @@
-#include "../graphics.hpp"
+#include "../inc/graphics.hpp"
 #include <SDL3/SDL_video.h>
 #include <cstdlib>
 
@@ -40,7 +40,7 @@ float *Color::asArrayPtr() {
 /* ------------ Uniform ------------ */
 
 Uniform::Uniform(const String &name, uint32 shaderID) {
-  id = glGetUniformLocation(shaderID, name.c_str());
+  setID(glGetUniformLocation(shaderID, name.c_str()));
   // TODO: add opengl error check
 
   this->name = name;
@@ -48,7 +48,90 @@ Uniform::Uniform(const String &name, uint32 shaderID) {
 
 const String &Uniform::getName() const { return name; }
 
-uint32 Uniform::getID() const { return id; }
+/* ------------ VBO ------------ */
+
+// TODO: add error checking
+VBO::VBO(float *verts, size_t size) {
+  glGenBuffers(1, getID_Ptr());
+  // THROW_ERROR_GL(FATAL.Derived("GL_GEN_BUFFERS_FAIL",
+  //                              "Generating the VBO buffer failed!"));
+
+  glBindBuffer(GL_ARRAY_BUFFER, getID());
+  // THROW_ERROR_GL(FATAL.Derived("GL_BIND_BUFFER_FAIL",
+  //                              "Binding VBO failed (Constructor)!"));
+
+  glBufferData(GL_ARRAY_BUFFER, size, verts, GL_STATIC_DRAW);
+  // THROW_ERROR_GL(FATAL.Derived("GL_BUFFER_DATA_FAIL",
+  //                              "Assigning data to the VBO failed!"));
+}
+
+VBO::~VBO() { glDeleteBuffers(1, getID_Ptr()); }
+
+// TODO: add error checking
+void VBO::bind() {
+  glBindBuffer(GL_ARRAY_BUFFER, getID());
+  // THROW_ERROR_GL(FATAL.Derived("GL_BIND_BUFFER_FAIL", "Binding VBO
+  // failed!"));
+}
+
+void VBO::unbind() { glBindBuffer(GL_ARRAY_BUFFER, 0); }
+
+/* ------------ VAO ------------ */
+
+// TODO: add error checking
+VAO::VAO() {
+  glGenVertexArrays(1, getID_Ptr());
+  // THROW_ERROR_GL(FATAL.Derived("GL_GEN_VERTEX_ARRAYS_FAIL",
+  //                              "Generating the VAO buffer failed!"));
+}
+
+VAO::~VAO() { glDeleteVertexArrays(1, getID_Ptr()); }
+
+// TODO: add error checking
+void VAO::bind() {
+  glBindVertexArray(getID());
+  // THROW_ERROR_GL(FATAL.Derived("GL_BIND_VERTEX_ARRAY_FAIL",
+  //                              "Binding the VAO buffer failed!"));
+}
+
+void VAO::unbind() { glBindVertexArray(0); }
+
+// TODO: add error checking
+void VAO::linkAttrib(VBO &vbo, uint32 layout, uint32 components, uint32 type,
+                     size_t stride, void *offset) {
+  vbo.bind();
+  glVertexAttribPointer(layout, components, type, GL_FALSE, stride, offset);
+  // THROW_ERROR_GL(FATAL.Derived("GL_VERTEX_ATTRIB_POINTER_FAIL"));
+
+  glEnableVertexAttribArray(layout);
+  // THROW_ERROR_GL(FATAL.Derived("GL_ENABLE_VERTEX_ATTRIB_ARRAY_FAIL"));
+
+  vbo.unbind();
+}
+
+/* ------------ EBO ------------ */
+
+// TODO: add error checking
+EBO::EBO(uint32 *indices, size_t size) {
+  glGenBuffers(1, getID_Ptr());
+  // THROW_ERROR_GL(FATAL.Derived("GL_GEN_BUFFERS_FAIL",
+  //                              "Generating the EBO buffer failed!"));
+
+  glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, getID());
+  // THROW_ERROR_GL(
+  //     FATAL.Derived("GL_BIND_BUFFER_FAIL", "Binding the EBO buffer
+  //     failed!"));
+
+  glBufferData(GL_ELEMENT_ARRAY_BUFFER, size, indices, GL_STATIC_DRAW);
+  // THROW_ERROR_GL(FATAL.Derived("GL_BUFFER_DATA_FAIL",
+  //                              "Assigning data to the EBO buffer failed!"));
+}
+
+EBO::~EBO() { glDeleteBuffers(1, getID_Ptr()); }
+
+void EBO::bind() { glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, getID()); }
+
+void EBO::unbind() { glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0); }
 
 /* ------------ Renderer ------------ */