+++ engine/src/logger.cpp
@@ -1,15 +1,17 @@
-#include "../logger.hpp"
-#include "../memory.hpp"
+#include "../inc/logger.hpp"
+#include "../inc/memory.hpp"
 
 using namespace vge;
 
 void Logger::init() {
-  if (File::exists("game.log") == false)
+  if (File::exists(".log") == false)
     return;
 
-  if (File::fileSize("game.log") > 3 * 1024 * 1024) {
+  if (File::fileSize(".log") > 3 * 1024 * 1024) {
     Logger::log("Log cleared.", __FILE__, __FUNCTION__, true);
   }
+
+  Logger::log("Logger initialized.", __FILE__, __FUNCTION__);
 }
 
 void Logger::log(const String &msg, const String &file, const String &function,