+++ engine/vge.hpp
@@ -1,0 +1,19 @@
+#ifndef VGE_HPP
+#define VGE_HPP
+
+#include "inc/asset.hpp"
+#include "inc/common.hpp"
+#include "inc/engine.hpp"
+#include "inc/file.hpp"
+#include "inc/graphics.hpp"
+#include "inc/list.hpp"
+#include "inc/logger.hpp"
+#include "inc/mathf.hpp"
+#include "inc/matrix4.hpp"
+#include "inc/memory.hpp"
+#include "inc/time.hpp"
+#include "inc/vector2.hpp"
+#include "inc/vector3.hpp"
+#include "inc/window.hpp"
+
+#endif