+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 06/08/26
+
+- added VAO, VBO, AND EBO
+- added `mbToBytes` function in `File` struct.
+
 ## 05/08/26
 
 - working on `Texture` struct